requirements.txt
torch==2.0.1
torchvision==0.15.2
albumentations==1.3.1
numpy==1.24.3
scikit-learn==1.3.0
tqdm==4.65.0
PyYAML==6.0
opencv-python==4.7.0.72
pandas==2.0.2